+++ src/peigs/src/c/pgexit.c
@@ -82,7 +82,7 @@ void pgexit( info, msg, proclist, nprocs, work )
   
   if ( m != 0 ) {
     *info = -51;
-    fprintf( stderr, " %s  me = %d exiting via pgexit. \n", msg, mxmynd_() );
+    fprintf( stderr, " %s  me = %d exiting via pgexit. \n", msg, (int)mxmynd_() );
     xstop_( info );
   }
